requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-rest-knox==5.0.2
djangorestframework==3.15.2
Faker==37.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.2
python-decouple>=3.8
python-dotenv==1.0.0
PyMySQL==1.1.1
